pyproject.toml
[project]
name = "xiaohui-agentic-playground"
version = "0.1.0"
description = "A playground for agentic experiments"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.75.0",
    "datasets>=4.4.1",
    "ipykernel>=7.1.0",
    "openai>=2.14.0",
    "openai-agents>=0.6.3",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    # LangGraph dependencies
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
]
